alexa-old3/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
chromadb==0.4.18
sentence-transformers==2.2.2
openai==1.3.9
requests==2.31.0
python-multipart==0.0.6
